Fix #4: More exception detail on DefineState and AddTransition.
Now the 'default' value of a 'value type' is allowed as a stateId. New method bool AddTransition(...GetState...) is introduced to handle dynamic transition if 'default' is used as a state id. Old void AddTransition(...Func...) is still can be used for reference types and in case if 'default' value means 'no state'.